Refactor: socket 폴더명 stroke로 변환
