+++ participants/routing.py
@@ -1,11 +1,8 @@
 from django.urls import path
 
-from participants.personal_socket import event_consumers, group_consumers
-from participants.team_socket import event_consumers, group_consumers
+from participants.stroke import stroke_event_consumers, stroke_group_consumers
 
 websocket_urlpatterns = [
-    path('ws/participants/<int:participant_id>/event', event_consumers.EventParticipantConsumer.as_asgi()),
-    path('ws/participants/<int:participant_id>/group', group_consumers.GroupParticipantConsumer.as_asgi()),
-    path('ws/participants/<int:participant_id>/event/team', event_consumers.EventParticipantConsumer.as_asgi()),
-    path('ws/participants/<int:participant_id>/group/team', group_consumers.GroupParticipantConsumer.as_asgi()),
+    path('ws/stroke/participants/<int:participant_id>/event/stroke', stroke_event_consumers.EventParticipantConsumer.as_asgi()),
+    path('ws/stroke/participants/<int:participant_id>/group/stroke', stroke_group_consumers.GroupParticipantConsumer.as_asgi()),
 ]
