+++ participants/stroke/redis_interface.py
